renamed tab bar components and completed/improved functionality
