+++ bsg-frontend/apps/extension/pages/_app.tsx
@@ -1,12 +1,13 @@
+import { useState } from 'react';
 import '../../../packages/ui-styles/global.css';
 import type { AppProps } from 'next/app';
 import '@bsg/ui-styles/global.css';
 import { Poppins } from 'next/font/google';
 import DefaultPopup from './defaultPopup';
-import { TabBar } from '@bsg/components/tabBar';
-import { Sidebar } from '@bsg/components/sidebar';
+import { HeaderBar } from '@bsg/components/tabBar/headerBar';
+import { Sidebar } from '@bsg/components/tabBar/sidebar';
 import { useIsCollapsed } from '@/hooks/useIsCollapsed';
-import { useState } from 'react';
+import { useIsOverflowed } from '@/hooks/useIsOverflowed';
 
 const poppins = Poppins({ weight: '400', subsets: ['latin'] });
 
@@ -17,7 +18,6 @@ export default function App({ Component, pageProps }: AppProps) {
   const [isHovered, setIsHovered] = useState(false);
 
   // Redirect popup render
-
   if (isDefaultPopup) {
     return (
       <div className={poppins.className}>
@@ -27,20 +27,25 @@ export default function App({ Component, pageProps }: AppProps) {
   }
 
   // On Leetcode extension render
-
-  return isCollapsed ? (
-    <Sidebar />
-  ) : (
+  return (
     //<div className={poppins.className}>
     <div
       onMouseEnter={() => setIsHovered(true)}
       onMouseLeave={() => setIsHovered(false)}
+      onClick={() => {console.log("panel clicked")}}
       className="flex flex-col h-screen"
     >
-      <TabBar isHovered={isHovered} isInRoom={true} />
-      <div className="flex-1 overflow-auto">
-        <Component {...pageProps} />
-      </div>
+      {
+        isCollapsed ?
+        <Sidebar isHovered={isHovered} isInRoom={true} />
+        :
+        <>
+          <HeaderBar isHovered={isHovered} isInRoom={true} />
+          <div className="flex-1 flex overflow-x-auto">
+            <Component {...pageProps} />
+          </div>
+        </>
+      }
     </div>
   );
 };
